Hook up ULI API, extract WS API to common project
