+++ check-digit/src/main/resources/application.conf
@@ -7,6 +7,7 @@ hmda {
     http {
       host = "0.0.0.0"
       port = 9091
+      timeout = 10
     }
   }
 }
