+++ common-api/src/main/scala/hmda/api/http/routes/BaseWsApi.scala
@@ -1,4 +1,4 @@
-package hmda.api.http
+package hmda.api.http.routes
 
 import java.net.InetAddress
 import java.time.Instant
